Divide the `PageComponents` file into  separate files.
