--- client/src/main/kotlin/io/spine/examples/shareaware/client/PageComponents.kt
@@ -1,225 +1,0 @@
-/*
- * Copyright 2023, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.examples.shareaware.client
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.loadSvgPainter
-import androidx.compose.ui.unit.Density
-import androidx.compose.ui.unit.dp
-import io.spine.examples.shareaware.MoneyCalculator
-import io.spine.examples.shareaware.client.MoneyExtensions.asReadableString
-import io.spine.examples.shareaware.client.StringExtensions.validateNumber
-import io.spine.examples.shareaware.client.payment.Popup
-import io.spine.examples.shareaware.client.payment.PopupConfig
-import io.spine.money.Money
-import io.spine.util.Exceptions
-import java.io.IOException
-import java.net.URL
-
-/**
- * Displays the card that shows difference between two `Money` objects.
- */
-@Composable
-public fun PriceDifferenceCard(actualPrice: Money, previousPrice: Money?) {
-    val color: Color
-    val price: String
-    if (null == previousPrice) {
-        color = MaterialTheme.colorScheme.surfaceVariant
-        price = "~0.0"
-    } else if (MoneyCalculator.isGreater(actualPrice, previousPrice)) {
-        color = MaterialTheme.colorScheme.surfaceVariant
-        price = "+${MoneyCalculator.subtract(actualPrice, previousPrice).asReadableString()}"
-    } else {
-        color = MaterialTheme.colorScheme.error
-        price = "-${MoneyCalculator.subtract(previousPrice, actualPrice).asReadableString()}"
-    }
-    Box(
-        modifier = Modifier
-            .background(color, MaterialTheme.shapes.extraSmall)
-            .padding(2.dp),
-    ) {
-        Text(
-            text = price,
-            style = MaterialTheme.typography.headlineSmall,
-            color = MaterialTheme.colorScheme.onPrimary
-        )
-    }
-}
-
-/**
- * [Scaffold] component that supports displaying the popup at the bottom of it.
- *
- * @param containerColor the color used for the background of this scaffold
- * @param modifier the `Modifier` to be applied to this scaffold
- * @param popupConfig configuration of the popup
- * @param content the content of the scaffold
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-public fun Scaffold(
-    containerColor: Color = MaterialTheme.colorScheme.surface,
-    modifier: Modifier = Modifier,
-    popupConfig: PopupConfig,
-    content: @Composable () -> Unit
-) {
-    Scaffold(
-        modifier = modifier,
-        containerColor = containerColor,
-        bottomBar = {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(40.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Popup(popupConfig)
-            }
-        }
-    ) {
-        content()
-    }
-}
-
-/**
- * Displays the search field.
- *
- * @param value the text to be shown in the search field
- * @param onChange the callback that is triggered when the search value changes
- */
-@Composable
-public fun SearchField(
-    value: String,
-    onChange: (String) -> Unit
-) {
-    Box(
-        modifier = Modifier.padding(top = 10.dp, bottom = 10.dp)
-    ) {
-        Input(
-            value = value,
-            onChange = onChange,
-            placeholder = "Search",
-            isError = false,
-            containerColor = MaterialTheme.colorScheme.secondary,
-            leadingIcon = {
-                Icon(
-                    imageVector = Icons.Default.Search,
-                    contentDescription = "Search",
-                    tint = MaterialTheme.colorScheme.onSecondary,
-                    modifier = Modifier
-                        .size(22.dp)
-                        .padding(end = 5.dp)
-                )
-            },
-            contentPadding = PaddingValues(
-                start = 5.dp,
-                end = 16.dp,
-                top = 2.dp,
-                bottom = 2.dp,
-            )
-        )
-    }
-}
-
-/**
- * Displays the input component accepting only the numeric values.
- *
- * @param placeholder the label to be displayed inside the input container
- * @param onChange the callback that is triggered when the input's value changes
- */
-@Composable
-public fun NumericInput(
-    placeholder: String,
-    onChange: (String) -> Unit
-) {
-    val input = remember { mutableStateOf("") }
-    val onChange: (String) -> Unit = {
-        if (it.validateNumber()) {
-            input.value = it
-            onChange(it)
-        }
-    }
-    Input(
-        value = input.value,
-        onChange = onChange,
-        placeholder = placeholder,
-        isError = false
-    )
-}
-
-/**
- * Draws an image from the network by URL.
- *
- * @param url the URL of the image to load
- * @param density density that will be used to set the intrinsic size of the image
- * @param contentDescription what the image represents
- * @param modifier modifier used to adjust layout
- * @param contentScale scale parameter used to determine the aspect ratio scaling
- */
-@Composable
-public fun Image(
-    url: String,
-    density: Density,
-    contentDescription: String,
-    modifier: Modifier = Modifier,
-    contentScale: ContentScale = ContentScale.Inside,
-) {
-    val image: Painter? = try {
-        URL(url).openStream().buffered().use { loadSvgPainter(it, density) }
-    } catch (e: IOException) {
-        throw Exceptions.illegalArgumentWithCauseOf(e)
-    }
-    if (image != null) {
-        Image(
-            painter = image,
-            contentDescription = contentDescription,
-            contentScale = contentScale,
-            modifier = modifier
-        )
-    }
-}
