+++ client/src/main/kotlin/io/spine/examples/shareaware/client/market/ShareProfileTab.kt
@@ -46,13 +46,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
-import io.spine.examples.shareaware.client.Image
+import io.spine.examples.shareaware.client.component.Image
 import io.spine.examples.shareaware.client.MoneyExtensions.asReadableString
-import io.spine.examples.shareaware.client.PriceDifferenceCard
+import io.spine.examples.shareaware.client.component.PriceDifferenceCard
 import io.spine.examples.shareaware.client.PrimaryButton
-import io.spine.examples.shareaware.client.Scaffold
+import io.spine.examples.shareaware.client.component.Scaffold
 import io.spine.examples.shareaware.client.market.MarketPageModel.PurchaseOperationModel
-import io.spine.examples.shareaware.client.payment.PopupConfig
+import io.spine.examples.shareaware.client.component.PopupConfig
 import io.spine.examples.shareaware.share.Share
 import io.spine.money.Money
 import kotlinx.coroutines.Dispatchers
