+++ client/src/main/kotlin/io/spine/examples/shareaware/client/market/SharesListTab.kt
@@ -58,8 +58,8 @@ import androidx.compose.ui.unit.dp
 import io.spine.examples.shareaware.ShareId
 import io.spine.examples.shareaware.client.ModifierExtensions.bottomBorder
 import io.spine.examples.shareaware.client.MoneyExtensions.asReadableString
-import io.spine.examples.shareaware.client.PriceDifferenceCard
-import io.spine.examples.shareaware.client.SearchField
+import io.spine.examples.shareaware.client.component.PriceDifferenceCard
+import io.spine.examples.shareaware.client.component.SearchField
 import io.spine.examples.shareaware.share.Share
 import io.spine.money.Money
