+++ client/src/main/kotlin/io/spine/examples/shareaware/client/wallet/WalletPage.kt
@@ -56,15 +56,15 @@ import io.spine.examples.shareaware.ReplenishmentId
 import io.spine.examples.shareaware.WithdrawalId
 import io.spine.examples.shareaware.client.DesktopClient
 import io.spine.examples.shareaware.client.EntitySubscription
-import io.spine.examples.shareaware.client.Input
+import io.spine.examples.shareaware.client.component.Input
 import io.spine.examples.shareaware.client.PrimaryButton
-import io.spine.examples.shareaware.client.Scaffold
+import io.spine.examples.shareaware.client.component.Scaffold
 import io.spine.examples.shareaware.client.StringExtensions.asIban
 import io.spine.examples.shareaware.client.StringExtensions.asUsd
 import io.spine.examples.shareaware.client.StringExtensions.validateIban
 import io.spine.examples.shareaware.client.StringExtensions.validateMoney
-import io.spine.examples.shareaware.client.payment.Dialog
-import io.spine.examples.shareaware.client.payment.PopupConfig
+import io.spine.examples.shareaware.client.component.Dialog
+import io.spine.examples.shareaware.client.component.PopupConfig
 import io.spine.examples.shareaware.paymentgateway.rejection.Rejections.MoneyCannotBeTransferredFromUser
 import io.spine.examples.shareaware.paymentgateway.rejection.Rejections.MoneyCannotBeTransferredToUser
 import io.spine.examples.shareaware.wallet.WalletBalance
